Always allow custom wildcards
Removed unnecessary restriction that custom wildcards are only possible when 'before' and 'after' are false/disabled.
This allows queries like:
Text LIKE '%This _s _ T%t%', when "This ?s ? T*t" was entered as search query, while 'before' and 'after' are true.
One or the other can be false, too