Implementing automatic calling of commonProcess().
This change allows it to configure the component to automatically trigger commonProcess() in the beforeFilter. I've set the default to false by default but I think it should be in fact set to true because I don't see any bad side effects it could have. Also I've made it so that instead of just activating it you can specify an action it should be triggered on and pass both params to the commmonProcess().